refactor(inspection-service): extract inspection-annotations sub-service
Move updateMediaAnnotations / saveAnnotation / setCroppedCover /
saveCroppedItemPhoto into InspectionAnnotationsService. Ownership checks and
photo-key validation call back through the facade (getInspection /
isInspectionPhotoKey). The facade delegates; moved code is byte-identical.
